+++ frontend/src/managers/input.ts
@@ -2,15 +2,15 @@ import { Editor } from "@/interop/editor";
 import { DialogState } from "@/providers/dialog";
 import { FullscreenState } from "@/providers/fullscreen";
 import { PortfolioState } from "@/providers/portfolio";
+import { makeKeyboardModifiersBitfield, textInputCleanup, getLatinKey } from "@/utilities/keyboard-entry";
 
 type EventName = keyof HTMLElementEventMap | keyof WindowEventHandlersEventMap | "modifyinputfield";
-interface EventListenerTarget {
+type EventListenerTarget = {
 	addEventListener: typeof window.addEventListener;
 	removeEventListener: typeof window.removeEventListener;
-}
+};
 
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export function createInputManager(editor: Editor, container: HTMLElement, dialog: DialogState, document: PortfolioState, fullscreen: FullscreenState) {
+export function createInputManager(editor: Editor, container: HTMLElement, dialog: DialogState, document: PortfolioState, fullscreen: FullscreenState): void {
 	// eslint-disable-next-line @typescript-eslint/no-explicit-any
 	const listeners: { target: EventListenerTarget; eventName: EventName; action: (event: any) => void; options?: boolean | AddEventListenerOptions }[] = [
 		{ target: window, eventName: "resize", action: (): void => onWindowResize(container) },
@@ -84,7 +84,7 @@ export function createInputManager(editor: Editor, container: HTMLElement, dialo
 
 		if (shouldRedirectKeyboardEventToBackend(e)) {
 			e.preventDefault();
-			const modifiers = makeModifiersBitfield(e);
+			const modifiers = makeKeyboardModifiersBitfield(e);
 			editor.instance.on_key_down(key, modifiers);
 			return;
 		}
@@ -100,7 +100,7 @@ export function createInputManager(editor: Editor, container: HTMLElement, dialo
 
 		if (shouldRedirectKeyboardEventToBackend(e)) {
 			e.preventDefault();
-			const modifiers = makeModifiersBitfield(e);
+			const modifiers = makeKeyboardModifiersBitfield(e);
 			editor.instance.on_key_up(key, modifiers);
 		}
 	};
@@ -118,7 +118,7 @@ export function createInputManager(editor: Editor, container: HTMLElement, dialo
 		const inFloatingMenu = e.target instanceof Element && e.target.closest("[data-floating-menu-content]");
 		if (!viewportPointerInteractionOngoing && inFloatingMenu) return;
 
-		const modifiers = makeModifiersBitfield(e);
+		const modifiers = makeKeyboardModifiersBitfield(e);
 		editor.instance.on_mouse_move(e.clientX, e.clientY, e.buttons, modifiers);
 	};
 
@@ -140,7 +140,7 @@ export function createInputManager(editor: Editor, container: HTMLElement, dialo
 		}
 
 		if (viewportPointerInteractionOngoing) {
-			const modifiers = makeModifiersBitfield(e);
+			const modifiers = makeKeyboardModifiersBitfield(e);
 			editor.instance.on_mouse_down(e.clientX, e.clientY, e.buttons, modifiers);
 		}
 	};
@@ -149,7 +149,7 @@ export function createInputManager(editor: Editor, container: HTMLElement, dialo
 		if (!e.buttons) viewportPointerInteractionOngoing = false;
 
 		if (!textInput) {
-			const modifiers = makeModifiersBitfield(e);
+			const modifiers = makeKeyboardModifiersBitfield(e);
 			editor.instance.on_mouse_up(e.clientX, e.clientY, e.buttons, modifiers);
 		}
 	};
@@ -158,7 +158,7 @@ export function createInputManager(editor: Editor, container: HTMLElement, dialo
 		if (!e.buttons) viewportPointerInteractionOngoing = false;
 
 		if (!textInput) {
-			const modifiers = makeModifiersBitfield(e);
+			const modifiers = makeKeyboardModifiersBitfield(e);
 			editor.instance.on_double_click(e.clientX, e.clientY, e.buttons, modifiers);
 		}
 	};
@@ -185,7 +185,7 @@ export function createInputManager(editor: Editor, container: HTMLElement, dialo
 
 		if (inCanvas) {
 			e.preventDefault();
-			const modifiers = makeModifiersBitfield(e);
+			const modifiers = makeKeyboardModifiersBitfield(e);
 			editor.instance.on_mouse_scroll(e.clientX, e.clientY, e.buttons, e.deltaX, e.deltaY, e.deltaZ, modifiers);
 		}
 	};
@@ -252,389 +252,8 @@ export function createInputManager(editor: Editor, container: HTMLElement, dialo
 	};
 
 	// Add event bindings for the lifetime of the application
-	const addListeners = (): void => {
-		listeners.forEach(({ target, eventName, action, options }) => target.addEventListener(eventName, action, options));
-	};
-
-	const removeListeners = (): void => {
-		listeners.forEach(({ target, eventName, action }) => target.removeEventListener(eventName, action));
-	};
+	listeners.forEach(({ target, eventName, action, options }) => target.addEventListener(eventName, action, options));
 
 	// Resize on creation
-	addListeners();
 	onWindowResize(container);
-
-	return { removeListeners };
 }
-
-export function makeModifiersBitfield(e: WheelEvent | PointerEvent | KeyboardEvent): number {
-	return Number(e.ctrlKey) | (Number(e.shiftKey) << 1) | (Number(e.altKey) << 2);
-}
-
-// Necessary because innerText puts an extra newline character at the end when the text is more than one line.
-export function textInputCleanup(text: string): string {
-	if (text[text.length - 1] === "\n") return text.slice(0, -1);
-	return text;
-}
-
-// This function is a naive, temporary solution to allow non-Latin keyboards to fall back on the physical QWERTY layout
-function getLatinKey(e: KeyboardEvent): string | null {
-	const key = e.key.toLowerCase();
-	const isPrintable = isKeyPrintable(e.key);
-
-	// Control (non-printable) characters are handled normally
-	if (!isPrintable) return key;
-
-	// These non-Latin characters should fall back to the Latin equivalent at the key location
-	const LAST_LATIN_UNICODE_CHAR = 0x024f;
-	if (key.length > 1 || key.charCodeAt(0) > LAST_LATIN_UNICODE_CHAR) return keyCodeToKey(e.code);
-
-	// Otherwise, ths is a printable Latin character
-	return e.key.toLowerCase();
-}
-
-function keyCodeToKey(code: string): string | null {
-	// Letters
-	if (code.match(/^Key[A-Z]$/)) return code.replace("Key", "").toLowerCase();
-
-	// Numbers
-	if (code.match(/^Digit[0-9]$/)) return code.replace("Digit", "");
-	if (code.match(/^Numpad[0-9]$/)) return code.replace("Numpad", "");
-
-	// Function keys
-	if (code.match(/^F[1-9]|F1[0-9]|F20$/)) return code.replace("F", "").toLowerCase();
-
-	// Other characters
-	if (MAPPING[code]) return MAPPING[code];
-
-	return null;
-}
-
-const MAPPING: Record<string, string> = {
-	BracketLeft: "[",
-	BracketRight: "]",
-	Backslash: "\\",
-	Slash: "/",
-	Period: ".",
-	Comma: ",",
-	Equal: "=",
-	Minus: "-",
-	Quote: "'",
-	Semicolon: ";",
-	NumpadEqual: "=",
-	NumpadDivide: "/",
-	NumpadMultiply: "*",
-	NumpadSubtract: "-",
-	NumpadAdd: "+",
-	NumpadDecimal: ".",
-};
-
-function isKeyPrintable(key: string): boolean {
-	return !ALL_PRINTABLE_KEYS.has(key);
-}
-
-const ALL_PRINTABLE_KEYS = new Set([
-	// Modifier
-	"Alt",
-	"AltGraph",
-	"CapsLock",
-	"Control",
-	"Fn",
-	"FnLock",
-	"Meta",
-	"NumLock",
-	"ScrollLock",
-	"Shift",
-	"Symbol",
-	"SymbolLock",
-	// Legacy modifier
-	"Hyper",
-	"Super",
-	// White space
-	"Enter",
-	"Tab",
-	// Navigation
-	"ArrowDown",
-	"ArrowLeft",
-	"ArrowRight",
-	"ArrowUp",
-	"End",
-	"Home",
-	"PageDown",
-	"PageUp",
-	// Editing
-	"Backspace",
-	"Clear",
-	"Copy",
-	"CrSel",
-	"Cut",
-	"Delete",
-	"EraseEof",
-	"ExSel",
-	"Insert",
-	"Paste",
-	"Redo",
-	"Undo",
-	// UI
-	"Accept",
-	"Again",
-	"Attn",
-	"Cancel",
-	"ContextMenu",
-	"Escape",
-	"Execute",
-	"Find",
-	"Help",
-	"Pause",
-	"Play",
-	"Props",
-	"Select",
-	"ZoomIn",
-	"ZoomOut",
-	// Device
-	"BrightnessDown",
-	"BrightnessUp",
-	"Eject",
-	"LogOff",
-	"Power",
-	"PowerOff",
-	"PrintScreen",
-	"Hibernate",
-	"Standby",
-	"WakeUp",
-	// IME composition keys
-	"AllCandidates",
-	"Alphanumeric",
-	"CodeInput",
-	"Compose",
-	"Convert",
-	"Dead",
-	"FinalMode",
-	"GroupFirst",
-	"GroupLast",
-	"GroupNext",
-	"GroupPrevious",
-	"ModeChange",
-	"NextCandidate",
-	"NonConvert",
-	"PreviousCandidate",
-	"Process",
-	"SingleCandidate",
-	// Korean-specific
-	"HangulMode",
-	"HanjaMode",
-	"JunjaMode",
-	// Japanese-specific
-	"Eisu",
-	"Hankaku",
-	"Hiragana",
-	"HiraganaKatakana",
-	"KanaMode",
-	"KanjiMode",
-	"Katakana",
-	"Romaji",
-	"Zenkaku",
-	"ZenkakuHankaku",
-	// Common function
-	"F1",
-	"F2",
-	"F3",
-	"F4",
-	"F5",
-	"F6",
-	"F7",
-	"F8",
-	"F9",
-	"F10",
-	"F11",
-	"F12",
-	"Soft1",
-	"Soft2",
-	"Soft3",
-	"Soft4",
-	// Multimedia
-	"ChannelDown",
-	"ChannelUp",
-	"Close",
-	"MailForward",
-	"MailReply",
-	"MailSend",
-	"MediaClose",
-	"MediaFastForward",
-	"MediaPause",
-	"MediaPlay",
-	"MediaPlayPause",
-	"MediaRecord",
-	"MediaRewind",
-	"MediaStop",
-	"MediaTrackNext",
-	"MediaTrackPrevious",
-	"New",
-	"Open",
-	"Print",
-	"Save",
-	"SpellCheck",
-	// Multimedia numpad
-	"Key11",
-	"Key12",
-	// Audio
-	"AudioBalanceLeft",
-	"AudioBalanceRight",
-	"AudioBassBoostDown",
-	"AudioBassBoostToggle",
-	"AudioBassBoostUp",
-	"AudioFaderFront",
-	"AudioFaderRear",
-	"AudioSurroundModeNext",
-	"AudioTrebleDown",
-	"AudioTrebleUp",
-	"AudioVolumeDown",
-	"AudioVolumeUp",
-	"AudioVolumeMute",
-	"MicrophoneToggle",
-	"MicrophoneVolumeDown",
-	"MicrophoneVolumeUp",
-	"MicrophoneVolumeMute",
-	// Speech
-	"SpeechCorrectionList",
-	"SpeechInputToggle",
-	// Application
-	"LaunchApplication1",
-	"LaunchApplication2",
-	"LaunchCalendar",
-	"LaunchContacts",
-	"LaunchMail",
-	"LaunchMediaPlayer",
-	"LaunchMusicPlayer",
-	"LaunchPhone",
-	"LaunchScreenSaver",
-	"LaunchSpreadsheet",
-	"LaunchWebBrowser",
-	"LaunchWebCam",
-	"LaunchWordProcessor",
-	// Browser
-	"BrowserBack",
-	"BrowserFavorites",
-	"BrowserForward",
-	"BrowserHome",
-	"BrowserRefresh",
-	"BrowserSearch",
-	"BrowserStop",
-	// Mobile phone
-	"AppSwitch",
-	"Call",
-	"Camera",
-	"CameraFocus",
-	"EndCall",
-	"GoBack",
-	"GoHome",
-	"HeadsetHook",
-	"LastNumberRedial",
-	"Notification",
-	"MannerMode",
-	"VoiceDial",
-	// TV
-	"TV",
-	"TV3DMode",
-	"TVAntennaCable",
-	"TVAudioDescription",
-	"TVAudioDescriptionMixDown",
-	"TVAudioDescriptionMixUp",
-	"TVContentsMenu",
-	"TVDataService",
-	"TVInput",
-	"TVInputComponent1",
-	"TVInputComponent2",
-	"TVInputComposite1",
-	"TVInputComposite2",
-	"TVInputHDMI1",
-	"TVInputHDMI2",
-	"TVInputHDMI3",
-	"TVInputHDMI4",
-	"TVInputVGA1",
-	"TVMediaContext",
-	"TVNetwork",
-	"TVNumberEntry",
-	"TVPower",
-	"TVRadioService",
-	"TVSatellite",
-	"TVSatelliteBS",
-	"TVSatelliteCS",
-	"TVSatelliteToggle",
-	"TVTerrestrialAnalog",
-	"TVTerrestrialDigital",
-	"TVTimer",
-	// Media controls
-	"AVRInput",
-	"AVRPower",
-	"ColorF0Red",
-	"ColorF1Green",
-	"ColorF2Yellow",
-	"ColorF3Blue",
-	"ColorF4Grey",
-	"ColorF5Brown",
-	"ClosedCaptionToggle",
-	"Dimmer",
-	"DisplaySwap",
-	"DVR",
-	"Exit",
-	"FavoriteClear0",
-	"FavoriteClear1",
-	"FavoriteClear2",
-	"FavoriteClear3",
-	"FavoriteRecall0",
-	"FavoriteRecall1",
-	"FavoriteRecall2",
-	"FavoriteRecall3",
-	"FavoriteStore0",
-	"FavoriteStore1",
-	"FavoriteStore2",
-	"FavoriteStore3",
-	"Guide",
-	"GuideNextDay",
-	"GuidePreviousDay",
-	"Info",
-	"InstantReplay",
-	"Link",
-	"ListProgram",
-	"LiveContent",
-	"Lock",
-	"MediaApps",
-	"MediaAudioTrack",
-	"MediaLast",
-	"MediaSkipBackward",
-	"MediaSkipForward",
-	"MediaStepBackward",
-	"MediaStepForward",
-	"MediaTopMenu",
-	"NavigateIn",
-	"NavigateNext",
-	"NavigateOut",
-	"NavigatePrevious",
-	"NextFavoriteChannel",
-	"NextUserProfile",
-	"OnDemand",
-	"Pairing",
-	"PinPDown",
-	"PinPMove",
-	"PinPToggle",
-	"PinPUp",
-	"PlaySpeedDown",
-	"PlaySpeedReset",
-	"PlaySpeedUp",
-	"RandomToggle",
-	"RcLowBattery",
-	"RecordSpeedNext",
-	"RfBypass",
-	"ScanChannelsToggle",
-	"ScreenModeNext",
-	"Settings",
-	"SplitScreenToggle",
-	"STBInput",
-	"STBPower",
-	"Subtitle",
-	"Teletext",
-	"VideoModeNext",
-	"Wink",
-	"ZoomToggle",
-]);
